Проверки доступа к редактированию и удалению объявлений в приложении animals_app

+++ animals_app/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 from .models import Animal
 from .forms import AnimalForm
@@ -24,6 +26,7 @@ def view_detail_animal(request, animal_id):
     }
     return render(request, "animals/detail_animals.html", context=context)
 
+@login_required(login_url='users:login')
 def add_animal_view(request):
     if request.method == 'POST':
         form = AnimalForm(request.POST, request.FILES)
@@ -36,8 +39,13 @@ def add_animal_view(request):
         form = AnimalForm()
     return render(request, 'animals/add_animal.html', {'form': form})
 
+@login_required(login_url='users:login')
 def edit_animal_view(request, animal_id):
     animal = get_object_or_404(Animal, id=animal_id)
+    if animal.submit != request.user:
+        raise PermissionDenied(
+            "У вас нет прав на редактирование данного объявления."
+    )
     if request.method == 'POST':
         form = AnimalForm(request.POST, request.FILES, instance=animal)
         if form.is_valid():
@@ -47,9 +55,13 @@ def edit_animal_view(request, animal_id):
         form = AnimalForm(instance=animal)
     return render(request, 'animals/add_animal.html', {'form': form})
 
+@login_required(login_url='users:login')
 def delete_animal_view(request, animal_id):
     animal = get_object_or_404(Animal, id=animal_id)
-    print(animal)
+    if animal.submit != request.user:
+        raise PermissionDenied(
+            "У вас нет прав на редактирование данного объявления."
+    )
     if request.method == 'POST':
         animal.delete()
         return redirect('animals:index')
